+++ spec/features/config_definition_spec.rb
@@ -331,8 +331,8 @@ describe 'Config definition' do
     expect(config.settings.database.user?).to eq(false)
     expect(config.settings.database.host?).to eq(true)
     expect(config.settings.enable_mocks?).to eq(true)
-    # setting roots does not have the predicate form
-    expect { config.settings.database? }.to raise_error(Qonfig::UnknownSettingError)
+    # setting roots always returns true
+    expect(config.settings.database?).to eq(true)
 
     # reconfigure and check again
     config.configure do |conf|
@@ -345,7 +345,17 @@ describe 'Config definition' do
     expect(config.settings.database.user?).to eq(true)
     expect(config.settings.database.host?).to eq(false)
     expect(config.settings.enable_mocks?).to eq(false)
-    # setting roots does not have the predicate form
-    expect { config.settings.database? }.to raise_error(Qonfig::UnknownSettingError)
+    # setting roots always returns true
+    expect(config.settings.database?).to eq(true)
+
+    # clear all options
+    config.configure do |conf|
+      conf.database.user = nil
+      conf.database.host = nil
+      conf.enable_mocks = nil
+    end
+
+    # setting roots always returns true
+    expect(config.settings.database?).to eq(true)
   end
 end